Move Disks/TemporaryFileOnDisk -> Common/TemporaryFile
