+++ src/Common/TemporaryFile.h
