+++ src/Storages/MergeTree/DataPartStorageOnDisk.cpp
@@ -1,7 +1,7 @@
 #include <Storages/MergeTree/DataPartStorageOnDisk.h>
 #include <Storages/MergeTree/MergeTreeDataPartChecksum.h>
 #include <Disks/IVolume.h>
-#include <Disks/TemporaryFileOnDisk.h>
+#include <Common/TemporaryFile.h>
 #include <IO/WriteBufferFromFileBase.h>
 #include <IO/ReadBufferFromFileBase.h>
 #include <IO/ReadHelpers.h>
