+++ src/Storages/StorageLog.cpp
@@ -30,7 +30,7 @@
 #include <Backups/BackupEntryFromSmallFile.h>
 #include <Backups/IBackup.h>
 #include <Backups/RestorerFromBackup.h>
-#include <Disks/TemporaryFileOnDisk.h>
+#include <Common/TemporaryFile.h>
 
 #include <cassert>
 #include <chrono>
